Add 'replace_files' setting

+++ decrypt_kdm.py
@@ -5,10 +5,25 @@ import struct
 import sys
 
 
+def to_stderr(message, *args, **kws):
+    message = message.format(*args, **kws)
+    print >> sys.stderr, message
+
+
+def to_stdout(message, *args, **kws):
+    message.format(*args, **kws)
+    print message.format(*args, **kws)
+
+
 class Decrypter(object):
-    def __init__(self, output_extension='txt', numbers_delimiter='\t'):
+    def __init__(self,
+                 output_extension='txt',
+                 numbers_delimiter='\t',
+                 replace_files=False):
+
         self.output_extension = output_extension
         self.numbers_delimiter = numbers_delimiter
+        self.replace_files = replace_files
 
     def decrypted_lines(self, path):
         source = open(path, 'rb')
@@ -42,13 +57,18 @@ class Decrypter(object):
         return os.path.join(directory, new_name)
 
     def decrypt(self, filepath):
-        print 'Decrypting {}...'.format(filepath)
-
         output_path = self.get_output_path(filepath)
+        if not self.replace_files and os.path.isfile(output_path):
+            to_stderr('WARNING: file {} elready exists', output_path)
+            to_stdout('Skip decrypting of {}\n', filepath)
+            return
+
+        to_stdout('Decrypting {}...', filepath)
+
         with open(output_path, 'w') as output:
             output.writelines(self.decrypted_lines(filepath))
 
-        print 'Decrypted file: {}\n'.format(output_path)
+        to_stdout('Decrypted file: {}\n', output_path)
 
 
 def main():
@@ -57,11 +77,17 @@ def main():
     if not filepaths:
         raise SystemExit('Nothing to decrypt')
 
+    # Set input arguments to change decrypter behaviour; for example:
+    # decrypter = Decrypter(
+    #     numbers_delimiter=' ',
+    #     output_extension='dcr',
+    #     replace_files=True
+    # )
     decrypter = Decrypter()
 
     for filepath in filepaths:
         if not os.path.isfile(filepath):
-            print >> sys.stderr, 'WARNING: file {!r} wasn\'t found\n'.format(filepath)
+            to_stderr('WARNING: file {} wasn\'t found\n', filepath)
             continue
 
         decrypter.decrypt(filepath)
